confAndMapAlmostDone

+++ include/validate.h
@@ -1,17 +1,17 @@
 #ifndef VALIDATE_H
 # define VALIDATE_H
 
+#include <stddef.h>
+
 typedef	struct s_vRGB
 {
-	int8_t	comma;
-	int8_t	digit;
+	size_t	comma;
+	size_t	digit;
 	bool	charac;
 }	t_vRGB;
 
 int		validate_RGB(char *str);
-
 int		validate_RES(char *str);
-
 int		validate_TEX(char *str);
 
 #endif
